fhir-pipeline/requirements.txt
requests==2.31.0
pandas==2.2.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
pytest==8.0.0
